% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{addselfloops}
\alias{addselfloops}
\title{addselfloops}
\usage{
addselfloops(data, A, B)
}
\arguments{
\item{data}{nX3 matrix containing triplets (first 2 columns and the
random state (data[,3]). The data matrix should be normalized
so that 
the row and column indices are continous, i.e., there is no
node to which there are no links. A and B contain the lower
and upper bounds for the vertex weights, respectively.}
}
\value{
Returns self-loops and lower and upper bounds for their
    weights.
}
\description{
addselfloops
}
