% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{BFS}
\alias{BFS}
\title{Performs Breadth-first search for a tree and finds a spanning tree.}
\usage{
BFS(graph, root = sample.int(length(graph), size = 1),
  Q = makequeue(length(graph)), bft = rep(NA, length(graph)),
  depth = rep(NA, length(graph)), st = rep(NA, length(graph)),
  treecount = 1)
}
\arguments{
\item{graph}{a list whose elements correspond to nodes and contain
a vector of nodes neighbours.}

\item{root}{index of the root.}
}
\value{
A vector which contains the index of the parent node. The
    root node's parent node has a index of zero.
}
\description{
Performs Breadth-first search for a tree and finds a spanning tree.
}
