% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bft2idx}
\alias{bft2idx}
\title{Finds the triplet indices that match the links in the spanning tree.}
\usage{
bft2idx(bft, triplets)
}
\arguments{
\item{bft}{spanning tree, e.g., output by BFS}

\item{triplets}{nX2 matrix that contains the links.}
}
\value{
Vector of length length(bft) that contains the triplet
    indices that match the links in the spanning tree.
}
\description{
Finds the triplet indices that match the links in the spanning tree.
}
