% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclesampler.R
\docType{package}
\name{cyclesampler-package}
\alias{cyclesampler}
\alias{cyclesampler-package}
\title{cyclesampler: Randomization Algorithms for Large Sparse Matrices}
\description{
A property-preserving Markov Chain Monte Carlo method for
	     generating surrogate networks in which (i) edge weights
	     are constrained to an interval and vertex weights are
	     preserved exactly, and (ii) edge and vertex weights are
	     both constrained to intervals.
}
\author{
\strong{Maintainer}: Kai Puolamäki \email{kai.puolamaki@iki.fi}

Other contributors:
\itemize{
  \item University of Helsinki [copyright holder]
}

}
\keyword{internal}
