% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cyclesampler}
\alias{cyclesampler}
\title{Cyclesampler object.}
\usage{
cyclesampler(data, a = rep(0, dim(data)[1]), b = rep(1, dim(data)[1]),
  useC = TRUE, nukezeros = TRUE)
}
\arguments{
\item{data}{n X 3 matrix containing triplets (first 2 columns and
the random state (data[, 3]). The data matrix should be
normalized so that the row and column indices are continous,
i.e., there is no node to which there are no links.}
}
\value{
Returns two function $getstate() that gives the current
    random state and $samplecycles(n) that samples n cycles.
}
\description{
Cyclesampler object.
}
