% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{delta1cycle}
\alias{delta1cycle}
\title{Finds a random term to add to the cycle.}
\usage{
delta1cycle(x)
}
\arguments{
\item{x}{random states in the cycle where the first half are the
"positive" and the second half "negative" entries.}
}
\value{
A random vector to number to add to the cycle.
}
\description{
Finds a random term to add to the cycle.
}
