% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{descaledata}
\alias{descaledata}
\title{Descales the values of data from (0,1) back to original format}
\usage{
descaledata(x, minv = 1, maxv = 5, drange = maxv - minv + 1,
  toint = TRUE)
}
\arguments{
\item{toint}{if TRUE values are rounded to next integers.}

\item{data}{data matrix where columns indicate row indices, column
indices and values, respectively}
}
\value{
Data matrix with value (third column) scaled to (0,1)
    interval.
}
\description{
Descales the values of data from (0,1) back to original format
}
