% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fexpU01}
\alias{fexpU01}
\title{expectation of exponential distribution defined in [0,1]}
\usage{
fexpU01(x, eps = .Machine$double.eps^0.25)
}
\arguments{
\item{x}{parameter of distribution in interval [-Inf,Inf]}
}
\value{
The expectation in interval [0,1].
}
\description{
expectation of exponential distribution defined in [0,1]
}
