% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{find1cycle}
\alias{find1cycle}
\title{Finds 1 cycle from the spanning tree.}
\usage{
find1cycle(bft, depth, idx, idx0, left, right)
}
\arguments{
\item{bft}{Spannig tree.}

\item{depth}{depth of the items in spanning tree.}

\item{idx}{Triplet indices of the links in spanning tree.}

\item{idx0}{Cycle link that does not appear in the spanning tree.}

\item{left}{triplets[idx0,1]}

\item{right}{triplets[idx0,2]}
}
\value{
Cycle extracted from the spanning tree.
}
\description{
Finds 1 cycle from the spanning tree.
}
