% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{findlambdaC}
\alias{findlambdaC}
\title{find values of lambdas in maxent method}
\usage{
findlambdaC(n, m, nu, u, s, l = rep(0, m), tol = 0.001, tole = -1,
  tolz = -1, maxiter = 1000, initlambda = TRUE, check = TRUE)
}
\arguments{
\item{n}{number of values}

\item{m}{number of constraints}

\item{nu}{number of constraints per value, vector of length n}

\item{u}{vector of constraints}

\item{s}{vector of expectations (e.g., row/column sums)}

\item{l}{initial values of lambdas (not used)}

\item{tol}{accuracy of lambdas in root finding}

\item{tole}{accuracy of expectations}

\item{tolz}{z-normalized accuracy of expectations}

\item{maxiter}{maximum number of iterations}

\item{check}{boolean, whether to do sanity checks for inputs}
}
\value{
list containing new lambdas, tolerance values (tol, tole,
    tolz) as well as numer of iterations done.
}
\description{
find values of lambdas in maxent method
}
