% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{findmerge}
\alias{findmerge}
\title{Finds node at which the tree defined by bft merges if starting from
nodes left and right.}
\usage{
findmerge(bft, depth, left, right)
}
\arguments{
\item{bft}{Spannig tree.}

\item{depth}{depth of the items in spanning tree.}

\item{left}{node}

\item{right}{node that is connected to left}
}
\value{
Node at which the trees with leafs at left and right merge.
}
\description{
Finds node at which the tree defined by bft merges if starting from
nodes left and right.
}
