% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{findrows}
\alias{findrows}
\title{Auxiliary function that indexes a vector.}
\usage{
findrows(rows)
}
\arguments{
\item{rows}{Vector of integers (e.g., row indices related to
values), containing values in [1,m] where m is the number of
constraints}
}
\value{
A list whose i'th element is a list containing the indices
    of values related to the i'th constraint.
}
\description{
Auxiliary function that indexes a vector.
}
