% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_node_weights}
\alias{get_node_weights}
\title{Get node weights}
\usage{
get_node_weights(data)
}
\arguments{
\item{data}{A 3-column matrix with the triplets.}
}
\value{
The Frobenius norm for each of the n_samples with respect to the starting state.
}
\description{
Get node weights
}
