% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makegraph}
\alias{makegraph}
\title{Makes a graph of triplets.}
\usage{
makegraph(triplets)
}
\arguments{
\item{triplets}{nX2 matrix containing the links.}
}
\value{
List where the i'th element is the list of nodes node i
    links to.
}
\description{
Makes a graph of triplets.
}
