% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makequeue}
\alias{makequeue}
\title{Efficient implementation of a standard FIFO queue using cyclic
array.}
\usage{
makequeue(size)
}
\arguments{
\item{size}{maximum size of the queue.}
}
\value{
Returns methods $isempty(), $push(x), and $eject().
}
\description{
Efficient implementation of a standard FIFO queue using cyclic
array.
}
