% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{maxentsampler}
\alias{maxentsampler}
\title{Maxentsampler2 object.}
\usage{
maxentsampler(data)
}
\arguments{
\item{data}{nX3 matrix containing triplets (first 2 columns and the
random state (data[,3]). The data matrix should be normalized
so that the random states are in the interval [0,1] and that
the row and column indices are continous, i.e., there is no
node to which there are no links.}
}
\value{
Returns functions $optimlambda() that finds the parameters
    lambda, $sample() that produces one sample, and $list() that gives
    a listing of internal parameters.
}
\description{
Maxentsampler2 object.
}
