% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nonzero}
\alias{nonzero}
\title{Finds mapping from indices so that they are mapped to continuous range
where each index has some instances. E.g., if s==c(1,3,4) we obtain
mapping p <- c(1,NA,2,3), i.e., p[s]==1:3.}
\usage{
nonzero(s, nbins = max(s), limit = 1)
}
\arguments{
\item{s}{List of indices (positive integers)}
}
\value{
Mapping as described above.
}
\description{
Finds mapping from indices so that they are mapped to continuous range
where each index has some instances. E.g., if s==c(1,3,4) we obtain
mapping p <- c(1,NA,2,3), i.e., p[s]==1:3.
}
