% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{samplecycles}
\alias{samplecycles}
\title{Samples n cycles and alters the random state accordingly}
\usage{
samplecycles(n, bft, depth, idx, idx0, triplets, randomstate)
}
\arguments{
\item{n}{Number of cycles to sample}

\item{bft}{Spannig tree.}

\item{depth}{depth of the items in spanning tree.}

\item{idx}{Triplet indices of the links in spanning tree.}

\item{idx0}{Vector of cycle links that do not appear in the spanning tree.}

\item{triplets}{nX2 matrix of triplest (links)}

\item{randomstate}{Current random state.}
}
\value{
New random state.
}
\description{
Samples n cycles and alters the random state accordingly
}
