% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{samplecycles2C}
\alias{samplecycles2C}
\title{C interface that is functionally equivalent to
samplecycles. Samples n cycles and alters the random state
accordingly}
\usage{
samplecycles2C(n, bft, depth, st, idx, idx0, slength, odds, triplets,
  randomstate, a = rep(0, dim(triplets)[1]), b = rep(1,
  dim(triplets)[1]), what = "randomstate")
}
\arguments{
\item{n}{Number of cycles to sample}

\item{bft}{Spannig tree.}

\item{depth}{depth of the items in spanning tree.}

\item{st}{subtree indices}

\item{idx}{Triplet indices of the links in spanning tree.}

\item{idx0}{Vector of cycle links that do not appear in the
spanning tree.}

\item{slength}{length of cycles corresponding to idx0}

\item{odds}{array with the number of odd cycles for each subgraph}

\item{triplets}{nX2 matrix of triplest (links)}

\item{randomstate}{Current random state.}

\item{a}{lower bound for the weight of each triplet}

\item{b}{upper bound for the weight of each triplet}

\item{what}{define what to return (string). One of \code{all} returning all
parameters, \code{randomstate} returning the new random state,
or \code{acceptance_rate} returning the fraction of accepted
moves.}
}
\value{
New random state.
}
\description{
C interface that is functionally equivalent to
samplecycles. Samples n cycles and alters the random state
accordingly
}
