% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sampleU01}
\alias{sampleU01}
\title{Samples of exponential distribution parametrized by x}
\usage{
sampleU01(x, n = length(x), y = runif(n),
  eps = .Machine$double.eps^0.25)
}
\arguments{
\item{x}{(A vector of) parameters to exponential distribution}
}
\value{
A vector of length length(x) of samples in interval [0,1].
}
\description{
Samples of exponential distribution parametrized by x
}
