% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scaledata}
\alias{scaledata}
\title{Scales the values of data to (0,1) interval}
\usage{
scaledata(x, minv = min(x), maxv = max(x), drange = maxv - minv + 1)
}
\arguments{
\item{x}{data vector of values}
}
\value{
Data vectorscaled to (0,1) interval.
}
\description{
Scales the values of data to (0,1) interval
}
