% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trtesplit}
\alias{trtesplit}
\title{Makes training/test data split}
\usage{
trtesplit(data, nte = round(0.014 * dim(data)[1]), limit = 1)
}
\arguments{
\item{data}{Triplets where columns indicate row indices, column
indices and values, respectively}
}
\value{
Training set and test set.
}
\description{
Makes training/test data split
}
